moved aside into its own component and some style improvements
