+++ src/components/aside/index.tsx
@@ -1,0 +1,2 @@
+import Aside from "./Aside";
+export default Aside
